adding batch write pk insertion test
